Re-enable multipartPolicy based formDataPolicy
